moved is_valid_ip and restart_service funcs to helpers.py
