+++ agents/agent.py
@@ -458,79 +458,6 @@ class ServerAgent(object):
         """
         return self._is_process_running() and self.is_ssh_service_active()
 
-    def restart_service(self, service, attempts=3):
-        """
-        Attempts to restart a system service with exponential backoff
-        if it is found to be inactive. Logs each attempt and result.
-
-        Parameters:
-            agent (ServerAgent): Agent instance used for logging. Must have
-            `logger` and `fallback_logger` attributes.
-            service (str): Name of the system service to restart (e.g., 'ssh').
-
-        Returns:
-            bool: True if the service was restarted, False otherwise.
-
-        Notes:
-            This function uses 'sudo systemctl restart <service>' and expects
-            that the agent has sufficient privileges to perform the operation.
-            Backoff strategy uses 2s, 4s, and 8s delays between attempts.
-        """
-
-        maybe_log_message(
-            '%s not active. Attempting restart...' % service,
-            self.logger,
-            fallback_logger=self.fallback_logger
-            )
-
-        for i in range(1, attempts+1):
-            try:
-                maybe_log_message(
-                    'Restarting %s (delay before restart: %s).' % (
-                        service,
-                        delay
-                        ),
-                    self.logger,
-                    fallback_logger=self.fallback_logger
-                    )
-
-                delay = 2**i
-                time.sleep(delay)
-
-                retcode = subprocess.call([
-                    'sudo',
-                    'systemctl',
-                    'restart',
-                    service
-                    ])
-
-                if retcode == 0:
-                    maybe_log_message(
-                        '%s service restarted successfully.' % service,
-                        self.logger,
-                        fallback_logger=self.fallback_logger,
-                        level=logging.INFO
-                        )
-                    return True  # If restsrting successful
-                else:
-                    maybe_log_message(
-                        '%s restart failed with code %s.' % (service, retcode),
-                        self.logger,
-                        fallback_logger=self.fallback_logger
-                        )
-
-            except Exception as restart_err:
-                maybe_log_message(
-                    'Error during %s service restart: %s' % (
-                        service,
-                        restart_err),
-                    self.logger,
-                    fallback_logger=self.fallback_logger,
-                    exc_info=True
-                    )
-                return False  # Stop after first fatal error
-        return False  # If restsrting failed
-
     @abc.abstractmethod
     def maybe_restart_service(self):
         pass
