+++ agents/utils/configtools.py
@@ -1,8 +1,5 @@
 import logging
-import re
-import subprocess
 import sys
-import time
 
 import ConfigParser
 import pkg_resources
@@ -101,12 +98,3 @@ def get_config_option(
             )
 
     return value or default
-
-
-def is_valid_ip(output):
-    """
-    Validate if the output is a correctly formatted IPv4 address.
-    Returns:
-        bool: True if the output is a valid IP address, False otherwise.
-    """
-    return re.match(r'^\d{1,3}(\.\d{1,3}){3}$', output.strip()) is not None
